remove autosubmit label
