+++ auto_submit/test/requests/check_pull_request_test.dart
@@ -10,12 +10,12 @@ import 'package:logging/logging.dart';
 import 'package:auto_submit/requests/check_pull_request.dart';
 import 'package:auto_submit/requests/check_pull_request_queries.dart';
 import 'package:github/github.dart';
-import 'package:meta/meta.dart';
 import 'package:mockito/mockito.dart';
 import 'package:test/test.dart';
 import 'package:graphql/client.dart' hide Request, Response;
 
 import '../utilities/mocks.dart';
+import '../utilities/utils.dart' hide createQueryResult;
 import './github_webhook_test_data.dart';
 import '../src/request_handling/fake_pubsub.dart';
 import '../src/request_handling/fake_authentication.dart';
@@ -474,112 +474,6 @@ void main() {
       assert(pubsub.messagesQueue.isEmpty);
     });
   });
-}
-
-enum ReviewState {
-  APPROVED,
-  CHANGES_REQUESTED,
-}
-
-enum MemberType {
-  OWNER,
-  MEMBER,
-  OTHER,
-}
-
-@immutable
-class PullRequestReviewHelper {
-  const PullRequestReviewHelper({
-    required this.authorName,
-    required this.state,
-    required this.memberType,
-  });
-
-  final String authorName;
-  final ReviewState state;
-  final MemberType memberType;
-}
-
-@immutable
-class StatusHelper {
-  const StatusHelper(this.name, this.state);
-
-  static const StatusHelper flutterBuildSuccess = StatusHelper('luci-flutter', 'SUCCESS');
-  static const StatusHelper flutterBuildFailure = StatusHelper('luci-flutter', 'FAILURE');
-  static const StatusHelper otherStatusFailure = StatusHelper('other status', 'FAILURE');
-
-  final String name;
-  final String state;
-}
-
-class PullRequestHelper {
-  PullRequestHelper({
-    this.author = 'author1',
-    this.prNumber = 0,
-    this.repo = 'flutter',
-    this.authorAssociation = 'MEMBER',
-    this.title = 'some_title',
-    this.reviews = const <PullRequestReviewHelper>[
-      PullRequestReviewHelper(authorName: 'member', state: ReviewState.APPROVED, memberType: MemberType.MEMBER)
-    ],
-    this.lastCommitHash = 'oid',
-    this.lastCommitStatuses = const <StatusHelper>[StatusHelper.flutterBuildSuccess],
-    this.lastCommitMessage = '',
-    this.dateTime,
-  });
-
-  final int prNumber;
-  final String repo;
-  final String author;
-  final String authorAssociation;
-  final List<PullRequestReviewHelper> reviews;
-  final String lastCommitHash;
-  List<StatusHelper>? lastCommitStatuses;
-  final String? lastCommitMessage;
-  final DateTime? dateTime;
-  final String title;
-
-  RepositorySlug get slug => RepositorySlug('flutter', repo);
-
-  Map<String, dynamic> toEntry() {
-    return <String, dynamic>{
-      'author': <String, dynamic>{'login': author},
-      'authorAssociation': authorAssociation,
-      'id': prNumber.toString(),
-      'title': title,
-      'reviews': <String, dynamic>{
-        'nodes': reviews.map((PullRequestReviewHelper review) {
-          return <String, dynamic>{
-            'author': <String, dynamic>{'login': review.authorName},
-            'authorAssociation': review.memberType.toString().replaceFirst('MemberType.', ''),
-            'state': review.state.toString().replaceFirst('ReviewState.', ''),
-          };
-        }).toList(),
-      },
-      'commits': <String, dynamic>{
-        'nodes': <dynamic>[
-          <String, dynamic>{
-            'commit': <String, dynamic>{
-              'oid': lastCommitHash,
-              'pushedDate': (dateTime ?? DateTime.now().add(const Duration(hours: -2))).toUtc().toIso8601String(),
-              'message': lastCommitMessage,
-              'status': <String, dynamic>{
-                'contexts': lastCommitStatuses != null
-                    ? lastCommitStatuses!.map((StatusHelper status) {
-                        return <String, dynamic>{
-                          'context': status.name,
-                          'state': status.state,
-                          'targetUrl': 'https://${status.name}',
-                        };
-                      }).toList()
-                    : <dynamic>[]
-              },
-            },
-          }
-        ],
-      },
-    };
-  }
 }
 
 QueryResult createQueryResult(PullRequestHelper pullRequest) {
